fix: go type float syntax (`float` does not exist)

+++ pkg/godef/common_go_types.go
@@ -17,7 +17,7 @@ var (
 		Syntax: "uint",
 	}
 	GoTypeFloat = GoTypePrimitive{
-		Syntax: "float",
+		Syntax: "float64",
 	}
 	GoTypeBool = GoTypePrimitive{
 		Syntax: "bool",
